+++ packages/jaspr/lib/src/dom/raw_text/raw_text_vm.dart
@@ -25,12 +25,12 @@ class _RawTextElement extends LeafRenderObjectElement {
 
   @override
   RenderObject createRenderObject() {
-    final parent = parentRenderObjectElement!.renderObject;
+    final parent = parentRenderObjectElement!.renderObject as RawableRenderObject;
     return parent.createChildRenderText((component as RawText).text, true);
   }
 
   @override
   void updateRenderObject(covariant RenderObject renderObject) {
-    (renderObject as RenderText).update((component as RawText).text, true);
+    (renderObject as RawableRenderText).update((component as RawText).text, true);
   }
 }